fix: rework legal contacts component on PF
